Merge branch 'master' of github.com:Kinza59/College-CMS

+++ backend/controllers/StdInquiryController.php
@@ -239,20 +239,24 @@ class StdInquiryController extends Controller
      */
     public function actionDelete($id)
     {
-        $request = Yii::$app->request;
-        $this->findModel($id)->delete();
+        if( Yii::$app->user->can('delete-inquiry')){
+            $request = Yii::$app->request;
+            $this->findModel($id)->delete();
 
-        if($request->isAjax){
-            /*
-            *   Process for ajax request
-            */
-            Yii::$app->response->format = Response::FORMAT_JSON;
-            return ['forceClose'=>true,'forceReload'=>'#crud-datatable-pjax'];
-        }else{
-            /*
-            *   Process for non-ajax request
-            */
-            return $this->redirect(['index']);
+            if($request->isAjax){
+                /*
+                *   Process for ajax request
+                */
+                Yii::$app->response->format = Response::FORMAT_JSON;
+                return ['forceClose'=>true,'forceReload'=>'#crud-datatable-pjax'];
+            }else{
+                /*
+                *   Process for non-ajax request
+                */
+                return $this->redirect(['index']);
+            }
+        } else {
+            throw new ForbiddenHttpException(403, 'You are not authorized to perform this action');
         }
